+++ oshdb/src/main/java/org/heigit/ohsome/oshdb/osm/OSMEntity.java
@@ -161,7 +161,7 @@ public abstract class OSMEntity {
   @Override
   public String toString() {
     return String.format("ID:%d V:+%d+ TS:%d CS:%d VIS:%s UID:%d TAGS:%S", getId(), getVersion(),
-        getTimestamp().getRawUnixTimestamp(), getChangesetId(), isVisible(), getUserId(),
+        getTimestamp().getEpochSecond(), getChangesetId(), isVisible(), getUserId(),
         Arrays.toString(getRawTags()));
   }
 }