+++ oshdb/src/main/java/org/heigit/ohsome/oshdb/osm/OSMNode.java
@@ -57,7 +57,7 @@ public class OSMNode extends OSMEntity implements Comparable<OSMNode>, Serializa
       c = Integer.compare(Math.abs(version), Math.abs(o.version));
     }
     if (c == 0) {
-      c = Long.compare(timestamp.getRawUnixTimestamp(), o.timestamp.getRawUnixTimestamp());
+      c = Long.compare(timestamp.getEpochSecond(), o.timestamp.getEpochSecond());
     }
     return c;
   }