+++ oshdb/src/test/java/org/heigit/ohsome/oshdb/osm/OSMNodeTest.java
@@ -119,7 +119,7 @@ public class OSMNodeTest {
   public void testGetTimestamp() {
     OSMNode instance = new OSMNode(1L, 1, new OSHDBTimestamp(1L), 1L, 1, new int[] {}, 1000000000L, 1000000000L);
     long expResult = 1L;
-    long result = instance.getTimestamp().getRawUnixTimestamp();
+    long result = instance.getTimestamp().getEpochSecond();
     assertEquals(expResult, result);
   }